Code re-organization
