+++ src/SimpleWallet/ParseArguments.h
@@ -15,27 +15,11 @@ You should have received a copy of the GNU General Public License
 along with this program. If not, see <http://www.gnu.org/licenses/>.
 */
 
-#include "version.h"
-#include "CryptoNoteConfig.h"
+#pragma once
 
-#include <algorithm>
-#include <iostream>
-#include <iomanip>
+#include <SimpleWallet/Types.h>
+
 #include <string>
-
-struct Config
-{
-    bool exit;
-
-    bool walletGiven;
-    bool passGiven;
-
-    std::string host;
-    int port;
-
-    std::string walletFile;
-    std::string walletPass;
-};
 
 char* getCmdOption(char ** begin, char ** end, const std::string & option);
