+++ src/SimpleWallet/PasswordContainer.cpp
@@ -15,7 +15,9 @@
 // You should have received a copy of the GNU Lesser General Public License
 // along with Bytecoin.  If not, see <http://www.gnu.org/licenses/>.
 
+//////////////////////////////
 #include "PasswordContainer.h"
+//////////////////////////////
 
 #include <iostream>
 #include <memory.h>
