+++ src/SimpleWallet/PasswordContainer.h
