+++ src/SimpleWallet/SimpleWallet.cpp
@@ -15,7 +15,38 @@ You should have received a copy of the GNU General Public License
 along with this program. If not, see <http://www.gnu.org/licenses/>.
 */
 
+//////////////////////////////////////
 #include <SimpleWallet/SimpleWallet.h>
+//////////////////////////////////////
+
+#include <boost/algorithm/string.hpp>
+#include <boost/thread/thread.hpp>
+
+#include <Common/SignalHandler.h>
+#include <Common/StringTools.h>
+
+#include <CryptoNoteCore/Currency.h>
+
+#include <Logging/FileLogger.h>
+#include <Logging/LoggerManager.h>
+
+#include <NodeRpcProxy/NodeRpcProxy.h>
+
+#include <SimpleWallet/Commands.h>
+#include <SimpleWallet/Fusion.h>
+#include <SimpleWallet/Open.h>
+#include <SimpleWallet/ParseArguments.h>
+#include <SimpleWallet/Sync.h>
+#include <SimpleWallet/Transfer.h>
+#include <SimpleWallet/Tools.h>
+
+#ifdef _WIN32
+/* Prevents windows.h redefining min/max which breaks compilation */
+#define NOMINMAX
+#include <windows.h>
+#endif
+
+#include "version.h"
 
 int main(int argc, char **argv)
 {
@@ -198,8 +229,7 @@ void run(CryptoNote::WalletGreen &wallet, CryptoNote::INode &node,
        received any money yet. */
     if (action != Generate)
     {
-        findNewTransactions(node, walletInfo);
-
+        syncWallet(node, walletInfo);
     }
     else
     {
@@ -247,441 +277,6 @@ Maybe<std::shared_ptr<WalletInfo>> handleAction(CryptoNote::WalletGreen &wallet,
     {
         throw std::runtime_error("Unimplemented action!");
     }
-}
-
-std::shared_ptr<WalletInfo> createViewWallet(CryptoNote::WalletGreen &wallet)
-{
-    Crypto::SecretKey privateViewKey = getPrivateKey("Private View Key: ");
-
-    CryptoNote::AccountPublicAddress publicKeys;
-    uint64_t prefix;
-
-    std::string address;
-
-    while (true)
-    {
-        std::cout << "Public TRTL address: ";
-
-        std::getline(std::cin, address);
-        boost::algorithm::trim(address);
-
-        if (address.length() != 99)
-        {
-            std::cout << WarningMsg("Address is wrong length!") << std::endl
-                      << "It should be 99 characters long, but it is "
-                      << address.length() << " characters long!" << std::endl;
-        }
-        else if (address.substr(0, 4) != "TRTL")
-        {
-            std::cout << WarningMsg("Invalid address! It should start with "
-                                    "\"TRTL\"!") << std::endl;
-        }
-        else if (!CryptoNote::parseAccountAddressString(prefix, publicKeys,
-                                                        address))
-        {
-            std::cout << WarningMsg("Failed to parse TRTL address! Ensure you "
-                                    "have entered it correctly.")
-                      << std::endl;
-        }
-        else
-        {
-            break;
-        }
-    }
-
-    std::string walletFileName = getNewWalletFileName();
-    std::string walletPass = getWalletPassword(true);
-
-    wallet.createViewWallet(walletFileName, walletPass, address,
-                            privateViewKey);
-
-    std::cout << InformationMsg("\nYour view wallet " + address 
-                              + " has been successfully imported!")
-              << std::endl << std::endl;
-
-    viewWalletMsg();
-
-    return std::make_shared<WalletInfo>(walletFileName, walletPass, 
-                                        address, true, wallet);
-}
-
-std::shared_ptr<WalletInfo> importWallet(CryptoNote::WalletGreen &wallet)
-{
-    Crypto::SecretKey privateSpendKey = getPrivateKey("Private Spend Key: ");
-    Crypto::SecretKey privateViewKey = getPrivateKey("Private View Key: ");
-    return importFromKeys(wallet, privateSpendKey, privateViewKey);
-}
-
-std::shared_ptr<WalletInfo> mnemonicImportWallet(CryptoNote::WalletGreen
-                                                 &wallet)
-{
-    std::string mnemonicPhrase;
-
-    Crypto::SecretKey privateSpendKey;
-    Crypto::SecretKey privateViewKey;
-
-    do
-    {
-        std::cout << "Mnemonic Phrase (25 words): ";
-        std::getline(std::cin, mnemonicPhrase);
-        boost::algorithm::trim(mnemonicPhrase);
-    }
-    while (!crypto::ElectrumWords::is_valid_mnemonic(mnemonicPhrase,
-                                                     privateSpendKey));
-
-    CryptoNote::AccountBase::generateViewFromSpend(privateSpendKey, 
-                                                   privateViewKey);
-
-    return importFromKeys(wallet, privateSpendKey, privateViewKey);
-}
-
-std::shared_ptr<WalletInfo> importFromKeys(CryptoNote::WalletGreen &wallet,
-                                           Crypto::SecretKey privateSpendKey, 
-                                           Crypto::SecretKey privateViewKey)
-{
-    std::string walletFileName = getNewWalletFileName();
-    std::string walletPass = getWalletPassword(true);
-
-    connectingMsg();
-
-    wallet.initializeWithViewKey(walletFileName, walletPass, privateViewKey);
-
-    std::string walletAddress = wallet.createAddress(privateSpendKey);
-
-    std::cout << InformationMsg("\nYour wallet " + walletAddress 
-                              + " has been successfully imported!")
-              << std::endl << std::endl;
-
-    return std::make_shared<WalletInfo>(walletFileName, walletPass, 
-                                        walletAddress, false, wallet);
-}
-
-std::shared_ptr<WalletInfo> generateWallet(CryptoNote::WalletGreen &wallet)
-{
-    std::string walletFileName = getNewWalletFileName();
-    std::string walletPass = getWalletPassword(true);
-
-    
-    CryptoNote::KeyPair spendKey;
-    Crypto::SecretKey privateViewKey;
-
-    Crypto::generate_keys(spendKey.publicKey, spendKey.secretKey);
-    CryptoNote::AccountBase::generateViewFromSpend(spendKey.secretKey,
-                                                   privateViewKey);
-
-    wallet.initializeWithViewKey(walletFileName, walletPass, privateViewKey);
-
-    std::string walletAddress = wallet.createAddress(spendKey.secretKey);
-
-    promptSaveKeys(wallet);
-
-    std::cout << WarningMsg("If you lose these your wallet cannot be recreated!")
-              << std::endl << std::endl;
-
-    return std::make_shared<WalletInfo>(walletFileName, walletPass,
-                                        walletAddress, false, wallet);
-}
-
-Maybe<std::shared_ptr<WalletInfo>> openWallet(CryptoNote::WalletGreen &wallet,
-                                              Config &config)
-{
-    std::string walletFileName = getExistingWalletFileName(config);
-
-    bool initial = true;
-
-    while (true)
-    {
-        std::string walletPass;
-
-        /* Only use the command line pass once, otherwise we will infinite
-           loop if it is incorrect */
-        if (initial && config.passGiven)
-        {
-            walletPass = config.walletPass;
-        }
-        else
-        {
-            walletPass = getWalletPassword(false);
-        }
-
-        initial = false;
-
-        connectingMsg();
-
-        try
-        {
-            wallet.load(walletFileName, walletPass);
-
-            std::string walletAddress = wallet.getAddress(0);
-            
-            Crypto::SecretKey privateSpendKey
-                = wallet.getAddressSpendKey(0).secretKey;
-
-            if (privateSpendKey == CryptoNote::NULL_SECRET_KEY)
-            {
-                std::cout << std::endl
-                          << InformationMsg("Your view only wallet "
-                                          + walletAddress
-                                          + " has been successfully opened!")
-                          << std::endl << std::endl;
-
-                viewWalletMsg();
-
-                return Just<std::shared_ptr<WalletInfo>>
-                           (std::make_shared<WalletInfo>(walletFileName,
-                                                         walletPass, 
-                                                         walletAddress,
-                                                         true, 
-                                                         wallet));
-            }
-            else
-            {
-                std::cout << std::endl
-                          << InformationMsg("Your wallet "
-                                          + walletAddress
-                                          + " has been successfully opened!")
-                          << std::endl << std::endl;
-
-                return Just<std::shared_ptr<WalletInfo>>
-                           (std::make_shared<WalletInfo>(walletFileName,
-                                                         walletPass, 
-                                                         walletAddress,
-                                                         false, 
-                                                         wallet));
-            }
- 
-            return Just<std::shared_ptr<WalletInfo>>
-                       (std::make_shared<WalletInfo>(walletFileName,
-                                                     walletPass, 
-                                                     walletAddress,
-                                                     false,
-                                                     wallet));
-        }
-        catch (const std::system_error& e)
-        {
-            std::string walletSuccessBadPwdMsg = 
-                "Restored view public key doesn't correspond to secret key: "
-                "The password is wrong";
-
-            std::string walletSuccessBadPwdMsg2 =
-                "Restored spend public key doesn't correspond to secret key: "
-                "The password is wrong";
-
-            std::string walletLegacyBadPwdMsg =
-                ": The password is wrong";
-
-            std::string alreadyOpenMsg =
-                "MemoryMappedFile::open: The process cannot access the file "
-                "because it is being used by another process.";
-
-            /* Wallet version from file not == 6 */
-            std::string notAWalletMsg =
-                "Unsupported wallet version: Wrong version";
-            
-            /* Got EOF before reading wallet version */
-            std::string notAWalletMsg2 =
-                "Failed to read wallet version: Wrong version";
-
-            std::string errorMsg = e.what();
-                
-            /* There are three different error messages depending upon if we're
-               opening a walletgreen or a walletlegacy wallet */
-            if (errorMsg == walletSuccessBadPwdMsg || 
-                errorMsg == walletSuccessBadPwdMsg2 ||
-                errorMsg == walletLegacyBadPwdMsg)
-            {
-                std::cout << WarningMsg("Incorrect password! Try again.")
-                          << std::endl;
-            }
-            /* The message actually has a \r\n on the end but i'd prefer to
-               keep just the raw string in the source so check the it starts
-               with instead */
-            else if (boost::starts_with(errorMsg, alreadyOpenMsg))
-            {
-                std::cout << WarningMsg("Could not open wallet! It is already "
-                                        "open in another process.")
-                          << std::endl
-                          << WarningMsg("Check with a task manager that you "
-                                        "don't have simplewallet open twice.")
-                          << std::endl
-                          << WarningMsg("Also check you don't have another "
-                                        "wallet program open, such as a GUI "
-                                        "wallet or walletd.")
-                          << std::endl << std::endl;
-
-                std::cout << "Returning to selection screen..." << std::endl
-                          << std::endl;
-
-                return Nothing<std::shared_ptr<WalletInfo>>();
-            }
-            else if (errorMsg == notAWalletMsg || errorMsg == notAWalletMsg2)
-            {
-                std::cout << WarningMsg("Could not open wallet file! It "
-                                        "doesn't appear to be a valid wallet!")
-                          << std::endl
-                          << WarningMsg("Ensure you are opening a wallet "
-                                        "file, and the file has not gotten "
-                                        "corrupted.")
-                          << std::endl
-                          << WarningMsg("Try reimporting via keys, and always "
-                                        "close simplewallet with the exit "
-                                        "command to prevent corruption.")
-                          << std::endl << std::endl;
-
-                std::cout << "Returning to selection screen..." << std::endl
-                          << std::endl;
-
-                return Nothing<std::shared_ptr<WalletInfo>>();
-            }
-            else
-            {
-                std::cout << "Unexpected error: " << errorMsg << std::endl;
-                std::cout << "Please report this error message and what "
-                          << "you did to cause it." << std::endl << std::endl;
-
-                std::cout << "Returning to selection screen..." << std::endl
-                          << std::endl;
-
-                return Nothing<std::shared_ptr<WalletInfo>>();
-            }
-        }
-    }
-}
-
-Crypto::SecretKey getPrivateKey(std::string msg)
-{
-    size_t privateKeyLen = 64;
-    size_t size;
-
-    std::string privateKeyString;
-    Crypto::Hash privateKeyHash;
-    Crypto::SecretKey privateKey;
-    Crypto::PublicKey publicKey;
-
-    while (true)
-    {
-        std::cout << msg;
-
-        std::getline(std::cin, privateKeyString);
-        boost::algorithm::trim(privateKeyString);
-
-        if (privateKeyString.length() != privateKeyLen)
-        {
-            std::cout << WarningMsg("Invalid private key, should be 64 "
-                                    "characters! Try again.") << std::endl;
-            continue;
-        }
-        else if (!Common::fromHex(privateKeyString, &privateKeyHash, 
-                  sizeof(privateKeyHash), size)
-                 || size != sizeof(privateKeyHash))
-        {
-            std::cout << WarningMsg("Invalid private key, failed to parse! "
-                                    "Ensure you entered it correctly.")
-                      << std::endl;
-            continue;
-        }
-
-        privateKey = *(struct Crypto::SecretKey *) &privateKeyHash;
-
-        /* Just used for verification purposes before we pass it to
-           walletgreen */
-        if (!Crypto::secret_key_to_public_key(privateKey, publicKey))
-        {
-            std::cout << "Invalid private key, failed to parse! Ensure "
-                         "you entered it correctly." << std::endl;
-            continue;
-        }
-
-        return privateKey;
-    }
-}
-
-std::string getExistingWalletFileName(Config &config)
-{
-    bool initial = true;
-
-    std::string walletName;
-
-    while (true)
-    {
-        /* Only use wallet file once in case it is incorrect */
-        if (config.walletGiven && initial)
-        {
-            walletName = config.walletFile;
-        }
-        else
-        {
-            std::cout << "What is the name of the wallet you want to open?: ";
-            std::getline(std::cin, walletName);
-        }
-
-        initial = false;
-
-        std::string walletFileName = walletName + ".wallet";
-
-        if (walletName == "")
-        {
-            std::cout << WarningMsg("Wallet name can't be blank! Try again.")
-                      << std::endl;
-        }
-        /* Allow people to enter wallet name with or without file extension */
-        else if (boost::filesystem::exists(walletName))
-        {
-            return walletName;
-        }
-        else if (boost::filesystem::exists(walletFileName))
-        {
-            return walletFileName;
-        }
-        else
-        {
-            std::cout << WarningMsg("A wallet with the filename ")
-                      << InformationMsg(walletName)
-                      << WarningMsg(" or ")
-                      << InformationMsg(walletFileName)
-                      << WarningMsg(" doesn't exist!")
-                      << std::endl
-                      << "Ensure you entered your wallet name correctly."
-                      << std::endl;
-        }
-    }
-}
-
-std::string getNewWalletFileName()
-{
-    std::string walletName;
-
-    while (true)
-    {
-        std::cout << "What would you like to call your new wallet?: ";
-        std::getline(std::cin, walletName);
-
-        std::string walletFileName = walletName + ".wallet";
-
-        if (boost::filesystem::exists(walletFileName))
-        {
-            std::cout << WarningMsg("A wallet with the filename " 
-                                  + walletFileName + " already exists!")
-                      << std::endl
-                      << "Try another name." << std::endl;
-        }
-        else if (walletName == "")
-        {
-            std::cout << WarningMsg("Wallet name can't be blank! Try again.")
-                      << std::endl;
-        }
-        else
-        {
-            return walletFileName;
-        }
-    }
-}
-
-std::string getWalletPassword(bool verifyPwd)
-{
-    Tools::PasswordContainer pwdContainer;
-    pwdContainer.read_password(verifyPwd);
-    return pwdContainer.password();
 }
 
 Action getAction(Config &config)
@@ -751,73 +346,6 @@ Action getAction(Config &config)
     }
 }
 
-void promptSaveKeys(CryptoNote::WalletGreen &wallet)
-{
-    std::cout << "Welcome to your new wallet, here is your payment address:"
-              << std::endl << InformationMsg(wallet.getAddress(0))
-              << std::endl << std::endl 
-              << "Please copy your secret keys and mnemonic seed and store "
-              << "them in a secure location: " << std::endl;
-
-    printPrivateKeys(wallet, false);
-
-    std::cout << std::endl;
-}
-
-void exportKeys(std::shared_ptr<WalletInfo> &walletInfo)
-{
-    confirmPassword(walletInfo->walletPass);
-    printPrivateKeys(walletInfo->wallet, walletInfo->viewWallet);
-}
-
-void printPrivateKeys(CryptoNote::WalletGreen &wallet, bool viewWallet)
-{
-    Crypto::SecretKey privateViewKey = wallet.getViewKey().secretKey;
-
-    if (viewWallet)
-    {
-        std::cout << SuccessMsg("Private view key:")
-                  << std::endl
-                  << SuccessMsg(Common::podToHex(privateViewKey))
-                  << std::endl;
-        return;
-    }
-
-    Crypto::SecretKey privateSpendKey = wallet.getAddressSpendKey(0).secretKey;
-
-    Crypto::SecretKey derivedPrivateViewKey;
-
-    CryptoNote::AccountBase::generateViewFromSpend(privateSpendKey,
-                                                   derivedPrivateViewKey);
-
-    bool deterministicPrivateKeys = derivedPrivateViewKey == privateViewKey;
-
-    std::cout << SuccessMsg("Private spend key:")
-              << std::endl
-              << SuccessMsg(Common::podToHex(privateSpendKey))
-              << std::endl
-              << std::endl
-              << SuccessMsg("Private view key:")
-              << std::endl
-              << SuccessMsg(Common::podToHex(privateViewKey))
-              << std::endl;
-
-    if (deterministicPrivateKeys)
-    {
-        std::string mnemonicSeed;
-
-        crypto::ElectrumWords::bytes_to_words(privateSpendKey, 
-                                              mnemonicSeed,
-                                              "English");
-
-        std::cout << std::endl
-                  << SuccessMsg("Mnemonic seed:")
-                  << std::endl
-                  << SuccessMsg(mnemonicSeed)
-                  << std::endl;
-    }
-}
-
 void welcomeMsg()
 {
     std::cout << "Use the "
@@ -867,6 +395,60 @@ std::string getInputAndDoWorkWhileIdle(std::shared_ptr<WalletInfo> &walletInfo)
            something, but enough that we're not starving the CPU */
         std::this_thread::sleep_for(std::chrono::milliseconds(50));
     }
+}
+
+bool shutdown(CryptoNote::WalletGreen &wallet, CryptoNote::INode &node,
+              bool &alreadyShuttingDown)
+{
+    if (alreadyShuttingDown)
+    {
+        std::cout << "Patience little turtle, we're already shutting down!" 
+                  << std::endl;
+        return false;
+    }
+    else
+    {
+        alreadyShuttingDown = true;
+        std::cout << InformationMsg("Saving wallet and shutting down, please "
+                                    "wait...") << std::endl;
+    }
+
+    bool finishedShutdown = false;
+
+    boost::thread timelyShutdown([&finishedShutdown]
+    {
+        auto startTime = std::chrono::system_clock::now();
+
+        /* Has shutdown finished? */
+        while (!finishedShutdown)
+        {
+            auto currentTime = std::chrono::system_clock::now();
+
+            /* If not, wait for a max of 20 seconds then force exit. */
+            if ((currentTime - startTime) > std::chrono::seconds(20))
+            {
+                std::cout << WarningMsg("Wallet took too long to save! "
+                                        "Force closing.") << std::endl
+                          << "Bye." << std::endl;
+                exit(0);
+            }
+
+            std::this_thread::sleep_for(std::chrono::seconds(1));
+        }
+    });
+
+    wallet.save();
+    wallet.shutdown();
+    node.shutdown();
+
+    finishedShutdown = true;
+
+    /* Wait for shutdown watcher to finish */
+    timelyShutdown.join();
+
+    std::cout << "Bye." << std::endl;
+
+    return true;
 }
 
 void inputLoop(std::shared_ptr<WalletInfo> &walletInfo, CryptoNote::INode &node)
@@ -972,607 +554,4 @@ void inputLoop(std::shared_ptr<WalletInfo> &walletInfo, CryptoNote::INode &node)
                       << std::endl;
         }
     }
-}
-
-void help(bool viewWallet)
-{
-    std::cout << "Available commands:" << std::endl
-              << SuccessMsg("help", 25)
-              << "List this help message" << std::endl
-              << SuccessMsg("reset", 25)
-              << "Discard cached data and recheck for transactions" << std::endl
-              << SuccessMsg("bc_height", 25)
-              << "Show the blockchain height" << std::endl
-              << SuccessMsg("balance", 25)
-              << "Display how much TRTL you have" << std::endl
-              << SuccessMsg("export_keys", 25)
-              << "Export your private keys" << std::endl
-              << SuccessMsg("address", 25)
-              << "Displays your payment address" << std::endl
-              << SuccessMsg("exit", 25)
-              << "Exit and save your wallet" << std::endl
-              << SuccessMsg("save", 25)
-              << "Save your wallet state" << std::endl
-              << SuccessMsg("incoming_transfers", 25)
-              << "Show incoming transfers" << std::endl;
-                  
-    if (viewWallet)
-    {
-        std::cout << InformationMsg("Please note you are using a view only "
-                                    "wallet, and so cannot transfer TRTL.")
-                  << std::endl;
-    }
-    else
-    {
-        std::cout << SuccessMsg("outgoing_transfers", 25)
-                  << "Show outgoing transfers" << std::endl
-                  << SuccessMsg("list_transfers", 25)
-                  << "Show all transfers" << std::endl
-                  << SuccessMsg("quick_optimize", 25)
-                  << "Quickly optimize your wallet to send large amounts"
-                  << std::endl
-                  << SuccessMsg("full_optimize", 25)
-                  << "Fully optimize your wallet to send large amounts"
-                  << std::endl
-                  << SuccessMsg("transfer", 25)
-                  << "Send TRTL to someone" << std::endl;
-    }
-}
-
-void balance(CryptoNote::INode &node, CryptoNote::WalletGreen &wallet,
-             bool viewWallet)
-{
-    uint64_t unconfirmedBalance = wallet.getPendingBalance();
-    uint64_t confirmedBalance = wallet.getActualBalance();
-    uint64_t totalBalance = unconfirmedBalance + confirmedBalance;
-
-    uint32_t localHeight = node.getLastLocalBlockHeight();
-    uint32_t remoteHeight = node.getLastKnownBlockHeight();
-    uint32_t walletHeight = wallet.getBlockCount();
-
-    std::cout << "Available balance: "
-              << SuccessMsg(formatAmount(confirmedBalance)) << std::endl
-              << "Locked (unconfirmed) balance: "
-              << WarningMsg(formatAmount(unconfirmedBalance))
-              << std::endl << "Total balance: "
-              << InformationMsg(formatAmount(totalBalance)) << std::endl;
-
-    if (viewWallet)
-    {
-        std::cout << std::endl 
-                  << InformationMsg("Please note that view only wallets "
-                                    "can only track incoming transactions, "
-                                    "and so your wallet balance may appear "
-                                    "inflated.") << std::endl;
-    }
-
-    if (localHeight < remoteHeight)
-    {
-        std::cout << std::endl
-                  << InformationMsg("Your daemon is not fully synced with "
-                                    "the network!")
-                  << std::endl << "Your balance may be incorrect until you "
-                  << "are fully synced!" << std::endl;
-    }
-    /* Small buffer because wallet height doesn't update instantly like node
-       height does */
-    else if (walletHeight + 1000 < remoteHeight)
-    {
-        std::cout << std::endl
-                  << InformationMsg("The blockchain is still being scanned for "
-                                    "your transactions.")
-                  << std::endl
-                  << "Balances might be incorrect whilst this is ongoing."
-                  << std::endl;
-    }
-}
-
-void blockchainHeight(CryptoNote::INode &node, CryptoNote::WalletGreen &wallet)
-{
-    uint32_t localHeight = node.getLastLocalBlockHeight();
-    uint32_t remoteHeight = node.getLastKnownBlockHeight();
-    uint32_t walletHeight = wallet.getBlockCount();
-
-    /* This is the height that the wallet has been scanned to. The blockchain
-       can be fully updated, but we have to walk the chain to find our
-       transactions, and this number indicates that progress. */
-    std::cout << "Wallet blockchain height: ";
-
-    /* Small buffer because wallet height doesn't update instantly like node
-       height does */
-    if (walletHeight + 1000 > remoteHeight)
-    {
-        std::cout << SuccessMsg(std::to_string(walletHeight));
-    }
-    else
-    {
-        std::cout << WarningMsg(std::to_string(walletHeight));
-    }
-
-    std::cout << std::endl << "Local blockchain height: ";
-
-    if (localHeight == remoteHeight)
-    {
-        std::cout << SuccessMsg(std::to_string(localHeight));
-    }
-    else
-    {
-        std::cout << WarningMsg(std::to_string(localHeight));
-    }
-
-    std::cout << std::endl << "Network blockchain height: "
-              << SuccessMsg(std::to_string(remoteHeight)) << std::endl;
-
-    if (localHeight == 0 && remoteHeight == 0)
-    {
-        std::cout << WarningMsg("Uh oh, it looks like you don't have "
-                                "TurtleCoind open!")
-                  << std::endl;
-    }
-    else if (walletHeight + 1000 < remoteHeight && localHeight == remoteHeight)
-    {
-        std::cout << InformationMsg("You are synced with the network, but the "
-                                    "blockchain is still being scanned for "
-                                    "your transactions.")
-                  << std::endl
-                  << "Balances might be incorrect whilst this is ongoing."
-                  << std::endl;
-    }
-    else if (localHeight == remoteHeight)
-    {
-        std::cout << SuccessMsg("Yay! You are synced!") << std::endl;
-    }
-    else
-    {
-        std::cout << WarningMsg("Be patient, you are still syncing with the "
-                                "network!") << std::endl;
-    }
-}
-
-bool shutdown(CryptoNote::WalletGreen &wallet, CryptoNote::INode &node,
-              bool &alreadyShuttingDown)
-{
-    if (alreadyShuttingDown)
-    {
-        std::cout << "Patience little turtle, we're already shutting down!" 
-                  << std::endl;
-        return false;
-    }
-    else
-    {
-        alreadyShuttingDown = true;
-        std::cout << InformationMsg("Saving wallet and shutting down, please "
-                                    "wait...") << std::endl;
-    }
-
-    bool finishedShutdown = false;
-
-    boost::thread timelyShutdown([&finishedShutdown]
-    {
-        auto startTime = std::chrono::system_clock::now();
-
-        /* Has shutdown finished? */
-        while (!finishedShutdown)
-        {
-            auto currentTime = std::chrono::system_clock::now();
-
-            /* If not, wait for a max of 20 seconds then force exit. */
-            if ((currentTime - startTime) > std::chrono::seconds(20))
-            {
-                std::cout << WarningMsg("Wallet took too long to save! "
-                                        "Force closing.") << std::endl
-                          << "Bye." << std::endl;
-                exit(0);
-            }
-
-            std::this_thread::sleep_for(std::chrono::seconds(1));
-        }
-    });
-
-    wallet.save();
-    wallet.shutdown();
-    node.shutdown();
-
-    finishedShutdown = true;
-
-    /* Wait for shutdown watcher to finish */
-    timelyShutdown.join();
-
-    std::cout << "Bye." << std::endl;
-
-    return true;
-}
-
-CryptoNote::BlockDetails getBlock(uint32_t blockHeight,
-                                  CryptoNote::INode &node)
-{
-    CryptoNote::BlockDetails block;
-
-    /* No connection to turtlecoind */
-    if (node.getLastKnownBlockHeight() == 0)
-    {
-        return block;
-    }
-
-    std::promise<std::error_code> errorPromise;
-
-    auto e = errorPromise.get_future();
-
-    auto callback = [&errorPromise](std::error_code e)
-    {
-        errorPromise.set_value(e);
-    };
-
-    node.getBlock(blockHeight, block, callback);
-
-    if (e.get())
-    {
-        /* Prevent the compiler optimizing it out... */
-        std::cout << "";
-    }
-
-    return block;
-}
-
-std::string getBlockTime(CryptoNote::BlockDetails b)
-{
-    if (b.timestamp == 0)
-    {
-        return "";
-    }
-
-    std::time_t time = b.timestamp;
-    char buffer[100];
-    std::strftime(buffer, sizeof(buffer), "%F %R", std::localtime(&time));
-    return std::string(buffer);
-}
-
-void printOutgoingTransfer(CryptoNote::WalletTransaction t,
-                           CryptoNote::INode &node)
-{
-    std::string blockTime = getBlockTime(getBlock(t.blockHeight, node));
-
-    std::cout << WarningMsg("Outgoing transfer:")
-              << std::endl
-              << WarningMsg("Hash: " + Common::podToHex(t.hash))
-              << std::endl
-              << WarningMsg("Spent: " + formatAmount(-t.totalAmount - t.fee))
-              << std::endl
-              << WarningMsg("Fee: " + formatAmount(t.fee))
-              << std::endl
-              << WarningMsg("Total Spent: " + formatAmount(-t.totalAmount))
-              << std::endl;
-
-    std::string paymentID = getPaymentID(t.extra);
-
-    if (paymentID != "")
-    {
-        std::cout << WarningMsg("Payment ID: " + paymentID) << std::endl;
-    }
-
-    /* Couldn't get timestamp, maybe old node or turtlecoind closed */
-    if (blockTime != "")
-    {
-        std::cout << WarningMsg("Timestamp: " + blockTime) << std::endl;
-    }
-
-    std::cout << std::endl;
-}
-
-void printIncomingTransfer(CryptoNote::WalletTransaction t,
-                           CryptoNote::INode &node)
-{
-    std::string blockTime = getBlockTime(getBlock(t.blockHeight, node));
-
-    std::cout << SuccessMsg("Incoming transfer:")
-              << std::endl
-              << SuccessMsg("Hash: " + Common::podToHex(t.hash))
-              << std::endl
-              << SuccessMsg("Amount: " + formatAmount(t.totalAmount))
-              << std::endl;
-
-    std::string paymentID = getPaymentID(t.extra);
-
-    if (paymentID != "")
-    {
-        std::cout << SuccessMsg("Payment ID: " + paymentID) << std::endl;
-    }
-
-    /* Couldn't get timestamp, maybe old node or turtlecoind closed */
-    if (blockTime != "")
-    {
-        std::cout << SuccessMsg("Timestamp: " + blockTime) << std::endl;
-    }
-
-    std::cout << std::endl;
-}
-
-void listTransfers(bool incoming, bool outgoing, 
-                   CryptoNote::WalletGreen &wallet, CryptoNote::INode &node)
-{
-    size_t numTransactions = wallet.getTransactionCount();
-    int64_t totalSpent = 0;
-    int64_t totalReceived = 0;
-
-    for (size_t i = 0; i < numTransactions; i++)
-    {
-        CryptoNote::WalletTransaction t = wallet.getTransaction(i);
-
-        if (t.totalAmount < 0 && outgoing)
-        {
-            printOutgoingTransfer(t, node);
-            totalSpent += -t.totalAmount;
-        }
-        else if (t.totalAmount > 0 && incoming)
-        {
-            printIncomingTransfer(t, node);
-            totalReceived += t.totalAmount;
-        }
-    }
-
-    if (incoming)
-    {
-        std::cout << SuccessMsg("Total received: " 
-                              + formatAmount(totalReceived))
-                  << std::endl;
-    }
-
-    if (outgoing)
-    {
-        std::cout << WarningMsg("Total spent: " + formatAmount(totalSpent))
-                  << std::endl;
-    }
-}
-
-void checkForNewTransactions(std::shared_ptr<WalletInfo> &walletInfo)
-{
-    walletInfo->wallet.updateInternalCache();
-
-    size_t newTransactionCount = walletInfo->wallet.getTransactionCount();
-
-    if (newTransactionCount != walletInfo->knownTransactionCount)
-    {
-        for (size_t i = walletInfo->knownTransactionCount; 
-                    i < newTransactionCount; i++)
-        {
-            CryptoNote::WalletTransaction t 
-                = walletInfo->wallet.getTransaction(i);
-
-            /* Don't print outgoing or fusion transfers */
-            if (t.totalAmount > 0)
-            {
-                std::cout << std::endl
-                          << InformationMsg("New transaction found!")
-                          << std::endl
-                          << SuccessMsg("Incoming transfer:")
-                          << std::endl
-                          << SuccessMsg("Hash: " + Common::podToHex(t.hash))
-                          << std::endl
-                          << SuccessMsg("Amount: "
-                                      + formatAmount(t.totalAmount))
-                          << std::endl
-                          << getPrompt(walletInfo)
-                          << std::flush;
-            }
-        }
-
-        walletInfo->knownTransactionCount = newTransactionCount;
-    }
-}
-
-void reset(CryptoNote::INode &node, std::shared_ptr<WalletInfo> &walletInfo)
-{
-    std::cout << InformationMsg("Resetting wallet...") << std::endl;
-
-    walletInfo->knownTransactionCount = 0;
-
-    /* Wallet is now unitialized. You must reinit with load, initWithKeys,
-       or whatever. This function wipes the cache, then saves the wallet. */
-    walletInfo->wallet.clearCacheAndShutdown();
-
-    /* Now, we reopen the wallet. It now has no cached tx's, and balance */
-    walletInfo->wallet.load(walletInfo->walletFileName,
-                            walletInfo->walletPass);
-
-    /* Now we rescan the chain to re-discover our balance and transactions */
-    findNewTransactions(node, walletInfo);
-}
-
-void findNewTransactions(CryptoNote::INode &node, 
-                         std::shared_ptr<WalletInfo> &walletInfo)
-{
-    uint32_t localHeight = node.getLastLocalBlockHeight();
-    uint32_t walletHeight = walletInfo->wallet.getBlockCount();
-    uint32_t remoteHeight = node.getLastKnownBlockHeight();
-
-    size_t transactionCount = walletInfo->wallet.getTransactionCount();
-
-    int stuckCounter = 0;
-
-    if (localHeight != remoteHeight)
-    {
-        std::cout << "Your TurtleCoind isn't fully synced yet!" << std::endl
-                  << "Until you are fully synced, you won't be able to send "
-                  << "transactions,"
-                  << std::endl
-                  << "and your balance may be missing or "
-                  << "incorrect!" << std::endl << std::endl;
-    }
-
-    /* If we open a legacy wallet then it will load the transactions but not
-       have the walletHeight == transaction height. Lets just throw away the
-       transactions and rescan. */
-    if (walletHeight == 1 && transactionCount != 0)
-    {
-        std::cout << "Upgrading your wallet from an older version of the "
-                  << "software..." << std::endl << "Unfortunately, we have "
-                  << "to rescan the chain to find your transactions."
-                  << std::endl;
-        transactionCount = 0;
-        walletInfo->wallet.clearCaches(true, false);
-    }
-
-    if (walletHeight == 1)
-    {
-        std::cout << "Scanning through the blockchain to find transactions "
-                  << "that belong to you." << std::endl
-                  << "Please wait, this will take some time."
-                  << std::endl << std::endl;
-    }
-    else
-    {
-        std::cout << "Scanning through the blockchain to find any new "
-                  << "transactions you received"
-                  << std::endl
-                  << "whilst your wallet wasn't open."
-                  << std::endl
-                  << "Please wait, this may take some time."
-                  << std::endl << std::endl;
-    }
-
-    while (walletHeight < localHeight)
-    {
-        int counter = 1;
-
-        /* This MUST be called on the main thread! */
-        walletInfo->wallet.updateInternalCache();
-
-        localHeight = node.getLastLocalBlockHeight();
-        remoteHeight = node.getLastKnownBlockHeight();
-        std::cout << SuccessMsg(std::to_string(walletHeight))
-                  << " of " << InformationMsg(std::to_string(localHeight))
-                  << std::endl;
-
-        uint32_t tmpWalletHeight = walletInfo->wallet.getBlockCount();
-
-        int waitSeconds = 1;
-
-        /* Save periodically so if someone closes before completion they don't
-           lose all their progress */
-        if (counter % 60 == 0)
-        {
-            walletInfo->wallet.save();
-        }
-
-        if (tmpWalletHeight == walletHeight)
-        {
-            stuckCounter++;
-            waitSeconds = 3;
-
-            if (stuckCounter > 20)
-            {
-                std::string warning =
-                    "Syncing may be stuck. Try restarting Turtlecoind.\n"
-                    "If this persists, visit "
-                    "https://turtlecoin.lol/#contact for support.";
-                std::cout << WarningMsg(warning) << std::endl;
-            }
-            else if (stuckCounter > 19)
-            {
-                /*
-                   Calling save has the side-effect of starting
-                   and stopping blockchainSynchronizer, which seems
-                   to sometimes force the sync to resume properly.
-                   So we'll try this before warning the user.
-                */
-                std::cout << InformationMsg("Saving wallet.") << std::endl;
-                walletInfo->wallet.save();
-                waitSeconds = 5;
-            }
-        }
-        else
-        {
-            stuckCounter = 0;
-            walletHeight = tmpWalletHeight;
-
-            size_t tmpTransactionCount = walletInfo
-                                       ->wallet.getTransactionCount();
-
-            if (tmpTransactionCount != transactionCount)
-            {
-                for (size_t i = transactionCount; i < tmpTransactionCount; i++)
-                {
-                    CryptoNote::WalletTransaction t
-                        = walletInfo->wallet.getTransaction(i);
-
-                    /* Don't print out fusion transactions */
-                    if (t.totalAmount != 0)
-                    {
-                        std::cout << std::endl
-                                  << InformationMsg("New transaction found!")
-                                  << std::endl << std::endl;
-
-                        if (t.totalAmount < 0)
-                        {
-                            printOutgoingTransfer(t, node);
-                        }
-                        else
-                        {
-                            printIncomingTransfer(t, node);
-                        }
-                    }
-                }
-
-                transactionCount = tmpTransactionCount;
-            }
-        }
-
-        counter++;
-
-        std::this_thread::sleep_for(std::chrono::seconds(waitSeconds));
-    }
-
-    std::cout << std::endl
-              << SuccessMsg("Finished scanning blockchain!") << std::endl
-              << std::endl;
-
-    /* In case the user force closes, we don't want them to have to rescan
-       the whole chain. */
-    walletInfo->wallet.save();
-
-    walletInfo->knownTransactionCount = transactionCount;
-}
-
-ColouredMsg getPrompt(std::shared_ptr<WalletInfo> &walletInfo)
-{
-    const int promptLength = 20;
-    const std::string extension = ".wallet";
-
-    std::string walletName = walletInfo->walletFileName;
-
-    /* Filename ends in .wallet, remove extension */
-    if (std::equal(extension.rbegin(), extension.rend(), 
-                   walletInfo->walletFileName.rbegin()))
-    {
-        size_t extPos = walletInfo->walletFileName.find_last_of('.');
-
-        walletName = walletInfo->walletFileName.substr(0, extPos);
-    }
-
-    std::string shortName = walletName.substr(0, promptLength);
-
-    return InformationMsg("[TRTL " + shortName + "]: ");
-}
-
-void connectingMsg()
-{
-    std::cout << std::endl << "Making initial contact with TurtleCoind."
-              << std::endl
-              << "Please wait, this sometimes can take a long time..."
-              << std::endl << std::endl;
-}
-
-void viewWalletMsg()
-{
-    std::cout << InformationMsg("Please remember that when using a view wallet "
-                                "you can only view incoming transactions!")
-              << std::endl << "This means if you received 100 TRTL and then "
-              << "sent 50 TRTL, your balance would appear to still be 100 "
-              << "TRTL." << std::endl
-              << "To effectively use a view wallet, you should only deposit "
-              << "to this wallet." << std::endl
-              << "If you have since needed to withdraw, send your remaining "
-              << "balance to a new wallet, and import this as a new view "
-              << "wallet so your balance can be correctly observed."
-              << std::endl << std::endl;
 }
